# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Titan X — generated from openapi.yaml.
 * Do not edit by hand; run codegen in @workspace/api-spec.
 */

export interface AuthCredentials {
  email: string;
  /** @minLength 8 */
  password: string;
}