# lib/api-zod/src/generated/types/detectedObject.ts
/**
 * Titan X — generated from openapi.yaml.
 * Do not edit by hand; run codegen in @workspace/api-spec.
 */

export interface DetectedObject {
  label: string;
  confidence: number;
  bbox: number[];
}